refactor: clean up messageDelete event logic and enhance logging

+++ src/events/messageDelete.js
@@ -34,16 +34,16 @@ module.exports = {
 		try {
 			/* if (message.author.bot || message.channel.id !== '1272033491390828574')
 				return; */
-			
+
 			//logger.DBG(`Message deleted in [${message.channel.name}] by ${message.author.tag}: ${message.content}`);
 
-			//if (message.channelId === channels.CREACIONES) {
-				logger.DBG(require('util').inspect(message, { depth: null, colors: true, compact: false }))
-				/* if (message.hasThread) {
+			if (message.channelId === channels.CREACIONES) {
+				logger.DBG(message.hasThread)
+				if (message.hasThread) {
 					const thread = message.thread ?? await message.channel.threads.fetch(message.id);
 					await thread.delete();
-				} */
-			//}
+				}
+			}
 		} catch (error) {
 			logger.ERR(error);
 		}
